==> packages/vessel-manoeuvring-models/vessel_manoeuvring_models-0.0.3.tar.gz/vessel_manoeuvring_models-0.0.3/notebooks/15.49_EK_smoother_3DOF_thrust.ipynb <==
# %load imports.py
## Local packages:

%matplotlib inline
%load_ext autoreload
%autoreload 2
%config Completer.use_jedi = False  ## (To fix autocomplete)

## External packages:
import pandas as pd
pd.options.display.max_rows = 999
pd.options.display.max_columns = 999
pd.set_option("display.max_columns", None)
import numpy as np
np.set_printoptions(linewidth=150)

import numpy as np
import os
import matplotlib.pyplot as plt
#if os.name == 'nt':
#    plt.style.use('presentation.mplstyle')  # Windows

import plotly.express as px 
import plotly.graph_objects as go

import seaborn as sns
import sympy as sp
from sympy.physics.mechanics import (dynamicsymbols, ReferenceFrame,
                                      Particle, Point)
from sympy.physics.vector.printing import vpprint, vlatex
from IPython.display import display, Math, Latex
from vessel_manoeuvring_models.substitute_dynamic_symbols import run, lambdify

import pyro

import sklearn
import pykalman
from statsmodels.sandbox.regression.predstd import wls_prediction_std
import statsmodels.api as sm

from scipy.integrate import solve_ivp

## Local packages:
from vessel_manoeuvring_models.data import mdl

from vessel_manoeuvring_models.symbols import *
from vessel_manoeuvring_models.parameters import *
import vessel_manoeuvring_models.symbols as symbols
from vessel_manoeuvring_models import prime_system
from vessel_manoeuvring_models.models import regression
from vessel_manoeuvring_models.visualization.regression import show_pred
from vessel_manoeuvring_models.visualization.plot import track_plot

## Load models:
# (Uncomment these for faster loading):
import vessel_manoeuvring_models.models.vmm_martin  as vmm 
from vessel_manoeuvring_models.models.vmm import ModelSimulator
from example_ship2 import ship_parameters, df_parameters


if os.name == 'nt':
    plt.style.use('../docs/book/book.mplstyle')  # Windows
    
from vessel_manoeuvring_models.visualization.plot import track_plot, plot

from vessel_manoeuvring_models.extended_kalman_vmm import ExtendedKalman
from vessel_manoeuvring_models.extended_kalman_filter import loglikelihood

## Load test

id=22773
df, units, meta_data = mdl.load(dir_path = '../data/raw', id=id)
df.index = df.index.total_seconds()
df.index-=df.index[0]
df['x0']-=df.iloc[0]['x0']
df['y0']-=df.iloc[0]['y0']
df['psi']-=df.iloc[0]['psi']


fig,ax=plt.subplots()
fig.set_size_inches(10,10)
track_plot(df=df, lpp=meta_data.lpp, x_dataset='x0', y_dataset='y0',  psi_dataset='psi', beam=meta_data.beam, ax=ax);

## Filtering

### Lowpass filter

from numpy import cos as cos
from numpy import sin as sin
from vessel_manoeuvring_models.data.lowpass_filter import lowpass_filter

df_lowpass = df.copy()
t = df_lowpass.index
ts = np.mean(np.diff(t))
fs = 1/ts

position_keys = ['x0','y0','psi']
for key in position_keys:
    df_lowpass[key] = lowpass_filter(data=df_lowpass[key], fs=fs, cutoff=1, order=1)

df_lowpass['x01d_gradient'] = x1d_ = np.gradient(df_lowpass['x0'], t)
df_lowpass['y01d_gradient'] = y1d_ = np.gradient(df_lowpass['y0'], t)
df_lowpass['r'] = r_ = np.gradient(df_lowpass['psi'], t)

psi_ = df_lowpass['psi']

df_lowpass['u'] = x1d_*cos(psi_) + y1d_*sin(psi_)
df_lowpass['v'] = -x1d_*sin(psi_) + y1d_*cos(psi_)

velocity_keys = ['u','v','r']
for key in velocity_keys:
    df_lowpass[key] = lowpass_filter(data=df_lowpass[key], fs=fs, cutoff=1, order=1)

data = df.copy()
data['thrust'] = data['Prop/PS/Thrust'] + data['Prop/SB/Thrust']

data['u'] = df_lowpass['u']
data['v'] = df_lowpass['v']
data['r'] = df_lowpass['r']
data=data.iloc[200:-100]
data.index-=data.index[0]

### Extended Kalman filter

parameters = df_parameters["prime"].copy()
ek = ExtendedKalman(vmm=vmm, 
                    parameters=parameters, 
                    ship_parameters=ship_parameters)


#### Simulate with system model

E = np.array([
    [0,0,0],
    [0,0,0],
    [0,0,0],
    [1,0,0],
    [0,1,0],
    [0,0,1],
    ],
)

t = np.linspace(0,10,100)
data_ = pd.DataFrame(index=t)
data_['delta'] = 0.0
data_['thrust'] = 20
data_['x0'] = 0
data_['y0'] = 0
data_['psi'] = 0
data_['u'] = 2
data_['v'] = 0
data_['r'] = 0

ek.simulate(data=data, input_columns=['delta','thrust'], 
            E=E).tail()

data_frames = {'data':data, 'sim':ek.df_simulation}

fig,ax=plt.subplots()
styles = {
    'Mesurement' : {
        'linestyle' : '',
        'marker' : '.',
        'ms' : 1,
        'zorder':-10,
    },
    
    'Kalman filter' : {
        'lw' : 2,
    },
    
    
}

for label,df_ in data_frames.items():
    track_plot(
            df=df_,
            lpp=ship_parameters["L"],
            beam=ship_parameters["B"],
            ax=ax,
            label=label,
            plot_boats=True,
            **styles.get(label,{})
        );
ax.legend()

plot(data_frames, keys=ek.df_simulation.columns);

### Extended Kalman filter and RTS smoother

Cd = np.array([
    [1, 0, 0, 0, 0, 0],
    [0, 1, 0, 0, 0, 0],
    [0, 0, 1, 0, 0, 0],
])

E = np.array([
    [0,0,0],
    [0,0,0],
    [0,0,0],
    [1,0,0],
    [0,1,0],
    [0,0,1],
    ],
)

P_prd = np.diag([0.1, 0.1, np.deg2rad(0.01), 0.01, 0.01, np.deg2rad(0.01)])
Qd = np.diag([0.01, 0.01, np.deg2rad(0.01)])  #process variances: u,v,r

error_max_pos = 0.1
sigma_pos = error_max_pos/3
variance_pos = sigma_pos**2

error_max_psi = np.deg2rad(0.1)
sigma_psi = error_max_psi/3
variance_psi = sigma_psi**2

Rd = np.diag([variance_pos, variance_pos, variance_psi])

ek.filter(
        data=data, P_prd=P_prd, Qd=Qd, Rd=Rd, E=E, Cd=Cd, 
        input_columns=['delta','thrust'],
    )


ek.smoother();

dataframes = {
    'Mesurement' : data,
    #'Kalman filter' : ek.df_kalman,
    'RTS': ek.df_smooth,
}

styles = {
    'Mesurement' : {
        'linestyle' : '',
        'marker' : '.',
        'ms' : 1,
        'zorder':-10,
    },
    
    'Kalman filter' : {
        'lw' : 2,
    },
    
    
}

plot(dataframes = dataframes, 
     fig_size=(10,10), 
     styles = ['r-','g-','b-'],
     keys=['x0','y0','psi','u','v','r','u1d','v1d','r1d']);

ek2 = ek.copy()

loglikelihood(ek2.time_steps_smooth)

ek2.parameters

def vary(parameters):
    
    ek2.parameters = parameters
    
    ek2.filter(
        data=data, P_prd=P_prd, Qd=Qd, Rd=Rd, E=E, Cd=Cd, 
        input_columns=['delta','thrust'],
    )


    ek2.smoother();
    
    return ek2
    

from scipy.optimize import minimize


def _vary(x, keys:list):
    parameters = deepcopy(ek.parameters)
    parameters[keys] = x
    
    ek_ = vary(parameters)
    return ek_
    
def fun(x,keys:list):
    ek_ = _vary(x, keys)
    return -loglikelihood(ek_.time_steps_smooth)
        

from copy import deepcopy

eks = []
key = 'Ndelta'
keys = [key]
values = np.linspace(0.5*ek.parameters[key], 1.5*ek.parameters[key], 5)
for value in values:
        
    ek_ = _vary(value, keys)
    eks.append(ek_.copy())
    

loglikelihoods = []

for ek_ in eks:
    
    loglikelihoods.append(loglikelihood(ek_.time_steps_smooth))
    

fig,ax=plt.subplots()

ax.plot(values, loglikelihoods)
ax.set_xlabel(key)
ax.set_ylabel('loglikelihood')

ek.parameters.values

%%time
keys = ek.parameters.keys()
#keys = ['Ndelta']

keys = [
'Xu',
'Xrr',
'Ndelta',                           
'Nv',
'Nr',
]

parameters = ek.parameters[keys]
x0 = parameters.values
bounds = [(-np.inf,0.0001) if value < 0 else (0.0001,np.inf) for value in parameters.values]

#small=0.5
#large=3
#bounds = [(large*value,small*value) if value < 0 else (small*value,large*value) for value in parameters.values]

#res = minimize(fun, x0=x0, bounds=bounds, tol=10**6, args=(keys,))

res = minimize(fun, x0=x0, tol=10*4, args=(keys,), bounds=bounds, method='Nelder-Mead', options={'maxiter':200})

res

x0

ek_ = _vary(res.x, keys=keys)
ek_.simulate(data=data, input_columns=['delta','thrust'], 
            E=E).tail()

#keys = [
#'Xu',        
#'Xvr',          
#'Ydelta',       
#'Yv',           
#'Yur',          
#'Ndelta',                           
#'Nv',
#]
#
#x = [-0.003     , -0.006     ,  0.003     , -0.01231641,  0.00413222, -0.0015    , -0.00318395]
#ek_ = _vary(x, keys=keys)
#ek_.simulate(data=data, input_columns=['delta','thrust'], 
#            E=E, x0=data.iloc[0][["x0", "y0", "psi", "u", "v", "r"]].values).tail()

data_frames = {'data':data, 'sim':ek_.df_simulation}

fig,ax=plt.subplots()
styles = {
    'Mesurement' : {
        'linestyle' : '',
        'marker' : '.',
        'ms' : 1,
        'zorder':-10,
    },
    
    'Kalman filter' : {
        'lw' : 2,
    },
    
    
}

for label,df_ in data_frames.items():
    track_plot(
            df=df_,
            lpp=ship_parameters["L"],
            beam=ship_parameters["B"],
            ax=ax,
            label=label,
            plot_boats=True,
            **styles.get(label,{})
        );
ax.legend()

plot(data_frames, keys=ek.df_simulation.columns);

%%time
input = {'delta':0,'thrust':1.0}
psi = 0
u = 2
v = 0
r = 0
x_dot = run(
            ek._lambda_f,
            **ek.parameters,
            **ek.ship_parameters,
            **input,
            psi=psi,
            u=u,
            v=v,
            r=r,
        )

ek._lambda_f

parameters = deepcopy(ek.parameters)
ship_parameters = deepcopy(ek.ship_parameters)

from inspect import signature
s = signature(ek._lambda_f)
keys = list(set(ek.ship_parameters) & set(s.parameters.keys()))
ship_parameters = {key:value for key,value in ek.ship_parameters.items() if key in keys}
    


%%time

input = {'delta':0,'thrust':1.0}
psi = 0
u = 2
v = 0
r = 0

ek._lambda_f(**parameters,
             **input,
             **ship_parameters,
            psi=psi,
            u=u,
            v=v,
            r=r,)

Izz = 693.124396594905
L = 5.014563106796117
volume = 0.4410267843660858
mass = volume*1000
np.sqrt(Izz/mass)/L

206.6 = 5.014563106796117
volume = 30843
mass = volume*1000
np.sqrt(Izz/mass)/L

72.1/206.6

11.11/39.0
